Self-hosted embedding
Add a self-hosted section to the embedding guide which references
webrecorder/example-webarchive and webrecorder/web-replay-gen.
The introductory paragraph was tightened up a little bit in places. The
examples now use WACZ instead of WARC, but intro also mentions that
the replay-web-page component supports WARC. If this needs to be
switched back let me know!

Closes #114